+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 [Full changelog](https://github.com/mozilla/glean/compare/v42.3.2...main)
 
 * General
-  * Removed `invalid_timezone_offset` metric ([#](https://github.com/mozilla/glean/pull/))
+  * Removed `invalid_timezone_offset` metric ([#1923](https://github.com/mozilla/glean/pull/1923))
 
 # v42.3.2 (2021-12-15)